+++ source/convert_audio_to_libopus/changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.2</span>**
+- fix typo in ffmpeg lib import
+- add logic to map and copy any existing audio stream that is already encoded with opus
+
 **<span style="color:#56adda">0.0.1</span>**
 - initial release
 - based on convert_multichannel_audio_to_aac_or_ac3